+++ crates/wdl-engine/src/backend/lsf_apptainer.rs
@@ -65,6 +65,7 @@ use crate::backend::INITIAL_EXPECTED_NAMES;
 use crate::backend::TaskExecutionConstraints;
 use crate::backend::TaskExecutionResult;
 use crate::config::Config;
+use crate::config::DEFAULT_TASK_SHELL;
 use crate::config::LsfApptainerBackendConfig;
 use crate::config::TaskResourceLimitBehavior;
 use crate::http::Transferer;
@@ -822,8 +823,15 @@ impl TaskExecutionBackend for LsfApptainerBackend {
             }
         }
 
-        let container =
-            requirements::container(inputs, requirements, self.config.task.container.as_deref());
+        let container = requirements::container(
+            inputs,
+            requirements,
+            if self.config.task.container.is_empty() {
+                None
+            } else {
+                Some(&self.config.task.container)
+            },
+        );
         if let ContainerSource::Unknown(_) = &container {
             bail!("LSF Apptainer backend does not support unknown container source `{container:#}`")
         }
@@ -892,7 +900,11 @@ impl TaskExecutionBackend for LsfApptainerBackend {
                 .apptainer
                 .generate_script(
                     &backend_config.apptainer_config,
-                    self.config.task.shell.as_deref(),
+                    if self.config.task.shell.is_empty() {
+                        DEFAULT_TASK_SHELL
+                    } else {
+                        &self.config.task.shell
+                    },
                     &request,
                     self.cancellation.first(),
                 )