+++ crates/wdl-engine/src/backend/slurm_apptainer.rs
@@ -60,6 +60,7 @@ use crate::backend::INITIAL_EXPECTED_NAMES;
 use crate::backend::TaskExecutionConstraints;
 use crate::backend::TaskExecutionResult;
 use crate::config::Config;
+use crate::config::DEFAULT_TASK_SHELL;
 use crate::config::SlurmApptainerBackendConfig;
 use crate::config::TaskResourceLimitBehavior;
 use crate::http::Transferer;
@@ -923,8 +924,15 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
             }
         }
 
-        let container =
-            requirements::container(inputs, requirements, self.config.task.container.as_deref());
+        let container = requirements::container(
+            inputs,
+            requirements,
+            if self.config.task.container.is_empty() {
+                None
+            } else {
+                Some(&self.config.task.container)
+            },
+        );
         if let ContainerSource::Unknown(_) = &container {
             bail!(
                 "Slurm Apptainer backend does not support unknown container source `{container:#}`"
@@ -1000,7 +1008,11 @@ impl TaskExecutionBackend for SlurmApptainerBackend {
                 .apptainer
                 .generate_script(
                     &backend_config.apptainer_config,
-                    self.config.task.shell.as_deref(),
+                    if self.config.task.shell.is_empty() {
+                        DEFAULT_TASK_SHELL
+                    } else {
+                        &self.config.task.shell
+                    },
                     &request,
                     self.cancellation.first(),
                 )